refactor: 가독성 위해 HomeView,OnboardingView 구조분리 (#6)
